Move mocking helpers to their own module.
