+++ sandpiper/tests/helpers/__init__.py
@@ -1,2 +1,3 @@
 from . import discord
 from . import misc
+from . import mocking
